requirements.txt
streamlit==1.46.0
langchain==0.3.26
langgraph==0.4.8
sentence-transformers==4.1.0
qdrant-client==1.14.3
python-dotenv==1.1.1
langchain-google-genai==2.1.5
numpy==2.3.0
pandas==2.3.0
PyMuPDF==1.26.1
tqdm==4.67.1
